ai-dashboard/backend/requirements.txt
fastapi==0.136.0
uvicorn[standard]==0.45.0
duckdb==1.5.2
openai==2.32.0
pydantic-settings==2.14.0
python-dotenv==1.2.1
pytest==9.0.2
httpx==0.28.1
